Rename trie_rs modules map -> trie_map, str -> str_trie_map
Follow the std::collections convention of naming a module as the
snake_case of its main type (hash_map::HashMap). This also stops the
str module from shadowing the primitive type name in paths.

trie_rs::TrieMap, trie_rs::iter, and trie_rs::TrieCount re-exports are
unchanged. The public path trie_rs::str::StrTrieMap becomes
trie_rs::str_trie_map::StrTrieMap; term_suffix_index is its only
consumer. Test modules are renamed to match (str -> str_trie_map,
trie -> trie_map).
