+++ src/redisearch_rs/trie_rs/src/str_trie_map/iter/unfiltered.rs
@@ -13,7 +13,7 @@ use crate::{
 };
 
 /// Lexicographical-order iterator over a
-/// [`StrTrieMap`](crate::str::StrTrieMap).
+/// [`StrTrieMap`](crate::str_trie_map::StrTrieMap).
 ///
 /// See [`crate::iter::Iter`] for the underlying traversal.
 pub struct Iter<'a, Data>(iter::Iter<'a, Data, filter::VisitAll>);
@@ -32,7 +32,7 @@ impl<'a, Data> Iterator for Iter<'a, Data> {
     }
 }
 
-/// Decode a trie byte key back to a `String`. Keys enter the [`crate::str::StrTrieMap`]
+/// Decode a trie byte key back to a `String`. Keys enter the [`crate::str_trie_map::StrTrieMap`]
 /// exclusively via `&str` so they are UTF-8 by construction; the validating
 /// `from_utf8` call here is cheap and protects against any future raw-byte
 /// insertion at the lower layer.
