+++ src/redisearch_rs/trie_rs/src/trie_map/iter/lending_range.rs
